Fixed some -Wint-conversion warnings.
There are cases where the lv_obj_t is used to store an int value
rather than pointing to a piece of data.  Updated the existing
uses of this pattern with proper casts.

+++ src/menu_handler.c
@@ -8,7 +8,7 @@ lv_obj_t * menu_dispatch_table[DEVICE_MENU_VEC_MAX];
 
 void menu_dispatch(lv_event_t * e) {
     lv_obj_t * target = lv_event_get_target(e);
-    int menu_vec = lv_obj_get_user_data(target);
+    int menu_vec = (int)lv_obj_get_user_data(target);
     printf("Select submenu [%d]\n", menu_vec);
     lv_obj_fade_out(lv_layer_top(), 30, 0);
     lv_disp_load_scr(menu_dispatch_table[menu_vec]);
